requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
numpy>=1.20.0
pandas>=1.3.0
Pillow>=8.2.0
matplotlib>=3.4.0
tqdm>=4.61.0
facenet-pytorch